Revamp this project:
* Cleanup docker ignore file
* Add Dockerfile.dev and make_dev_env.sh to allow for creating dev environments more easily
* Point script dir to /app/src/script (better parity with existing folder structure)

One of the more annoying things about this project (currently) is the need to
rebuild the C++ code all the time. This normally wouldn't be an issue, but since
I need a live linux system (for modifying /etc/shadow and friends), I need to do
it in a docker container. This dev environment should help fix this.