Add simple DB migration program:
This will need to be added onto as time goes on, but it should be good enough
for now. This will basically need to be run on each update, and there will need
to be some sort of script to migrate the data to-from docker containers.

Also rename the build folder in the dev dockerfile to build-docker, so that you
don't have to wipe out the build folder when switching between local and dev
builds.